// models/Presentation.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 09 Dec 2021 17:29:41 GMT'
 */
export type Presentation = IContentItem<{
  media: Elements.LinkedItemsElement<IContentItem>;
  description: Elements.RichTextElement;
  topic: Elements.TaxonomyElement;
  name: Elements.TextElement;
}>;