// models/Room.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 09 Dec 2021 17:29:41 GMT'
 */
export type Room = IContentItem<{
  name: Elements.TextElement;
  floor: Elements.LinkedItemsElement<IContentItem>;
  journeyStage: Elements.TaxonomyElement;
  seatingCapacity: Elements.NumberElement;
  map: Elements.LinkedItemsElement<IContentItem>;
}>;