// models/Speaker.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 09 Dec 2021 17:29:42 GMT'
 */
export type Speaker = IContentItem<{
  bio: Elements.RichTextElement;
  media: Elements.LinkedItemsElement<IContentItem>;
  journeyStage: Elements.TaxonomyElement;
  firstName: Elements.TextElement;
  jobTitle: Elements.TextElement;
  lastName: Elements.TextElement;
}>;